+++ packages/animeflv/CHANGELOG.md
@@ -13,7 +13,8 @@
 
 ### Patch Changes
 
-- 46d5716: - Allow the user in `browseAnime()` and `browseFilter` to set the origin URL (desktop or mobile)
+- 46d5716: - Allow the user in `browseAnime()` and `browseFilter()` to set the origin URL (desktop
+  or mobile)
 
 ## 0.1.0
 
@@ -23,6 +24,8 @@
 
 ### Patch Changes
 
-- e3d7ba0: feat: getInformation it returns the information of the anime/title only for the desktop site.
-- 8190ed5: feat: getInformationMobile() it returns the information of the anime/title only for the mobile site.
+- e3d7ba0: feat: getInformation it returns the information of the anime/title only for the desktop
+  site.
+- 8190ed5: feat: getInformationMobile() it returns the information of the anime/title only for the
+  mobile site.
 - 54c025c: feat: add getPages() for cheerio